+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 2020-04-08
 
+### Added
+
+- Maintenance DAG with tasks to clean up old S3 files and temporary DB tables
+
 ### Changed
 
 - ONS UK Trade in services pipelines now include period types, and shouldn't have duplicated data for quarterly rows.
@@ -14,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Updated duty type field type to text on raw world bank tariff pipeline
 - Added columns eu_rep_rate and eu_part_rate to world bank tariff pipeline
 - Updated partner and reporter field types to text on world bank tariff pipeline
+- S3Data methods are now retried in case of temporary network failure
 
 ## 2020-04-07